+++ Cargo.toml
@@ -35,7 +35,7 @@ nu-table = { path = "./crates/nu-table" }
 nu-term-grid = { path = "./crates/nu-term-grid" }
 # nu-ansi-term = { path = "./crates/nu-ansi-term" }
 nu-ansi-term = "0.39.0"
-nu-config = { path = "./crates/nu-config" }
+nu-color-config = { path = "./crates/nu-color-config" }
 miette = "3.0.0"
 ctrlc = "3.2.1"
 crossterm_winapi = "0.9.0"
