+++ crates/nu-cli/Cargo.toml
@@ -10,7 +10,7 @@ nu-parser = { path = "../nu-parser" }
 nu-protocol = { path = "../nu-protocol" }
 # nu-ansi-term = { path = "../nu-ansi-term" }
 nu-ansi-term = "0.39.0"
-nu-config = { path = "../nu-config" }
+nu-color-config = { path = "../nu-color-config" }
 
 miette = { version = "3.0.0", features = ["fancy"] }
 thiserror = "1.0.29"
