+++ crates/nu-color-config/src/color_config.rs
@@ -1,7 +1,7 @@
+use crate::nu_style::{color_from_hex, color_string_to_nustyle};
 use nu_ansi_term::{Color, Style};
 use nu_protocol::Config;
 use nu_table::{Alignment, TextStyle};
-use serde::Deserialize;
 use std::collections::HashMap;
 
 //TODO: should this be implemented again?
@@ -14,105 +14,6 @@ use std::collections::HashMap;
 //         Number::Decimal(decimal) => Primitive::Decimal(decimal),
 //     }
 // }
-
-#[derive(Deserialize, PartialEq, Debug)]
-struct NuStyle {
-    fg: Option<String>,
-    bg: Option<String>,
-    attr: Option<String>,
-}
-
-fn parse_nustyle(nu_style: NuStyle) -> Style {
-    // get the nu_ansi_term::Color foreground color
-    let fg_color = match nu_style.fg {
-        Some(fg) => color_from_hex(&fg).expect("error with foreground color"),
-        _ => None,
-    };
-    // get the nu_ansi_term::Color background color
-    let bg_color = match nu_style.bg {
-        Some(bg) => color_from_hex(&bg).expect("error with background color"),
-        _ => None,
-    };
-    // get the attributes
-    let color_attr = match nu_style.attr {
-        Some(attr) => attr,
-        _ => "".to_string(),
-    };
-
-    // setup the attributes available in nu_ansi_term::Style
-    let mut bold = false;
-    let mut dimmed = false;
-    let mut italic = false;
-    let mut underline = false;
-    let mut blink = false;
-    let mut reverse = false;
-    let mut hidden = false;
-    let mut strikethrough = false;
-
-    // since we can combine styles like bold-italic, iterate through the chars
-    // and set the bools for later use in the nu_ansi_term::Style application
-    for ch in color_attr.to_lowercase().chars() {
-        match ch {
-            'l' => blink = true,
-            'b' => bold = true,
-            'd' => dimmed = true,
-            'h' => hidden = true,
-            'i' => italic = true,
-            'r' => reverse = true,
-            's' => strikethrough = true,
-            'u' => underline = true,
-            'n' => (),
-            _ => (),
-        }
-    }
-
-    // here's where we build the nu_ansi_term::Style
-    Style {
-        foreground: fg_color,
-        background: bg_color,
-        is_blink: blink,
-        is_bold: bold,
-        is_dimmed: dimmed,
-        is_hidden: hidden,
-        is_italic: italic,
-        is_reverse: reverse,
-        is_strikethrough: strikethrough,
-        is_underline: underline,
-    }
-}
-
-fn color_string_to_nustyle(color_string: String) -> Style {
-    // eprintln!("color_string: {}", &color_string);
-    if color_string.chars().count() < 1 {
-        Style::default()
-    } else {
-        let nu_style = match nu_json::from_str::<NuStyle>(&color_string) {
-            Ok(s) => s,
-            Err(_) => NuStyle {
-                fg: None,
-                bg: None,
-                attr: None,
-            },
-        };
-
-        parse_nustyle(nu_style)
-    }
-}
-
-fn color_from_hex(hex_color: &str) -> std::result::Result<Option<Color>, std::num::ParseIntError> {
-    // right now we only allow hex colors with hashtag and 6 characters
-    let trimmed = hex_color.trim_matches('#');
-    if trimmed.len() != 6 {
-        Ok(None)
-    } else {
-        // make a nu_ansi_term::Color::Rgb color by converting hex to decimal
-        Ok(Some(Color::Rgb(
-            u8::from_str_radix(&trimmed[..2], 16)?,
-            u8::from_str_radix(&trimmed[2..4], 16)?,
-            u8::from_str_radix(&trimmed[4..6], 16)?,
-        )))
-    }
-}
 
 pub fn lookup_ansi_color_style(s: String) -> Style {
     if s.starts_with('#') {
@@ -470,148 +371,6 @@ pub fn style_primitive(primitive: &str, color_hm: &HashMap<String, Style>) -> Te
         //     }
         // }
         _ => TextStyle::basic_left(),
-    }
-}
-
-pub fn get_shape_color(shape: String, conf: &Config) -> Style {
-    match shape.as_ref() {
-        "flatshape_garbage" => {
-            if conf.color_config.contains_key("flatshape_garbage") {
-                let int_color = &conf.color_config["flatshape_garbage"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::White).on(Color::Red).bold()
-            }
-        }
-        "flatshape_bool" => {
-            if conf.color_config.contains_key("flatshape_bool") {
-                let int_color = &conf.color_config["flatshape_bool"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::LightCyan)
-            }
-        }
-        "flatshape_int" => {
-            if conf.color_config.contains_key("flatshape_int") {
-                let int_color = &conf.color_config["flatshape_int"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Purple).bold()
-            }
-        }
-        "flatshape_float" => {
-            if conf.color_config.contains_key("flatshape_float") {
-                let int_color = &conf.color_config["flatshape_float"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Purple).bold()
-            }
-        }
-        "flatshape_range" => {
-            if conf.color_config.contains_key("flatshape_range") {
-                let int_color = &conf.color_config["flatshape_range"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Yellow).bold()
-            }
-        }
-        "flatshape_internalcall" => {
-            if conf.color_config.contains_key("flatshape_internalcall") {
-                let int_color = &conf.color_config["flatshape_internalcall"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Cyan).bold()
-            }
-        }
-        "flatshape_external" => {
-            if conf.color_config.contains_key("flatshape_external") {
-                let int_color = &conf.color_config["flatshape_external"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Cyan)
-            }
-        }
-        "flatshape_externalarg" => {
-            if conf.color_config.contains_key("flatshape_externalarg") {
-                let int_color = &conf.color_config["flatshape_externalarg"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Green).bold()
-            }
-        }
-        "flatshape_literal" => {
-            if conf.color_config.contains_key("flatshape_literal") {
-                let int_color = &conf.color_config["flatshape_literal"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Blue)
-            }
-        }
-        "flatshape_operator" => {
-            if conf.color_config.contains_key("flatshape_operator") {
-                let int_color = &conf.color_config["flatshape_operator"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Yellow)
-            }
-        }
-        "flatshape_signature" => {
-            if conf.color_config.contains_key("flatshape_signature") {
-                let int_color = &conf.color_config["flatshape_signature"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Green).bold()
-            }
-        }
-        "flatshape_string" => {
-            if conf.color_config.contains_key("flatshape_string") {
-                let int_color = &conf.color_config["flatshape_string"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Green)
-            }
-        }
-        "flatshape_filepath" => {
-            if conf.color_config.contains_key("flatshape_filepath") {
-                let int_color = &conf.color_config["flatshape_filepath"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Cyan)
-            }
-        }
-        "flatshape_globpattern" => {
-            if conf.color_config.contains_key("flatshape_globpattern") {
-                let int_color = &conf.color_config["flatshape_globpattern"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Cyan).bold()
-            }
-        }
-        "flatshape_variable" => {
-            if conf.color_config.contains_key("flatshape_variable") {
-                let int_color = &conf.color_config["flatshape_variable"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Purple)
-            }
-        }
-        "flatshape_flag" => {
-            if conf.color_config.contains_key("flatshape_flag") {
-                let int_color = &conf.color_config["flatshape_flag"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().fg(Color::Blue).bold()
-            }
-        }
-        "flatshape_custom" => {
-            if conf.color_config.contains_key("flatshape_custom") {
-                let int_color = &conf.color_config["flatshape_custom"];
-                lookup_ansi_color_style(int_color.to_string())
-            } else {
-                Style::new().bold()
-            }
-        }
-        _ => Style::default(),
     }
 }
