+++ crates/nu-command/Cargo.toml
@@ -17,7 +17,7 @@ nu-term-grid = { path = "../nu-term-grid" }
 nu-parser = { path = "../nu-parser" }
 # nu-ansi-term = { path = "../nu-ansi-term" }
 nu-ansi-term = "0.39.0"
-nu-config = { path = "../nu-config" }
+nu-color-config = { path = "../nu-color-config" }
 
 # Potential dependencies for extras
 url = "2.2.1"
