--- crates/nu-config/src/lib.rs
@@ -1,5 +1,0 @@
-mod color_config;
-
-pub use color_config::{
-    get_color_config, get_shape_color, lookup_ansi_color_style, style_primitive,
-};
